+++ libs/cgv_gl/surface_renderer.h
@@ -45,23 +45,23 @@ namespace cgv { // @<
 			void set_normal_array(const context& ctx, const std::vector<T>& normals) { has_normals = true;  set_attribute_array(ctx, ref_prog().get_attribute_location(ctx, "normal"), normals); }
 			/// templated method to set the normal attribute from an array of normals of type T, which should have 3 components
 			template <typename T>
-			void set_normal_array(const context& ctx, const T* normals, size_t nr_elements, size_t stride_in_bytes = 0) { has_normals = true;  set_attribute_array(ctx, ref_prog().get_attribute_location(ctx, "normal"), normals, nr_elements, stride_in_bytes); }
+			void set_normal_array(const context& ctx, const T* normals, size_t nr_elements, unsigned stride_in_bytes = 0) { has_normals = true;  set_attribute_array(ctx, ref_prog().get_attribute_location(ctx, "normal"), normals, nr_elements, stride_in_bytes); }
 			/// method to set the normal attribute from a vertex buffer object, the element type must be given as explicit template parameter
-			void set_normal_array(const context& ctx, type_descriptor element_type, const vertex_buffer& vbo, size_t offset_in_bytes, size_t nr_elements, size_t stride_in_bytes = 0);
+			void set_normal_array(const context& ctx, type_descriptor element_type, const vertex_buffer& vbo, size_t offset_in_bytes, size_t nr_elements, unsigned stride_in_bytes = 0);
 			/// template method to set the normal attribute from a vertex buffer object, the element type must be given as explicit template parameter
 			template <typename ElementType>
-			void set_normal_array(const context& ctx, const vertex_buffer& vbo, size_t offset_in_bytes, size_t nr_elements, size_t stride_in_bytes = 0) { set_normal_array(ctx, type_descriptor(element_descriptor_traits<T>::get_type_descriptor(T()), true), vbo, offset_in_bytes, nr_elements, stride_in_bytes); }
+			void set_normal_array(const context& ctx, const vertex_buffer& vbo, size_t offset_in_bytes, size_t nr_elements, unsigned stride_in_bytes = 0) { set_normal_array(ctx, type_descriptor(element_descriptor_traits<T>::get_type_descriptor(T()), true), vbo, offset_in_bytes, nr_elements, stride_in_bytes); }
 			/// templated method to set the texcoord attribute array from a vector of texcoords of type T
 			template <typename T>
 			void set_texcoord_array(const context& ctx, const std::vector<T>& texcoords) { has_texcoords = true;  set_attribute_array(ctx, ref_prog().get_attribute_location(ctx, "texcoord"), texcoords); }
 			/// templated method to set the texcoord attribute from an array of texcoords of type T
 			template <typename T>
-			void set_texcoord_array(const context& ctx, const T* texcoords, size_t nr_elements, size_t stride_in_bytes = 0) { has_texcoords = true;  set_attribute_array(ctx, ref_prog().get_attribute_location(ctx, "texcoord"), texcoords, nr_elements, stride_in_bytes); }
+			void set_texcoord_array(const context& ctx, const T* texcoords, size_t nr_elements, unsigned stride_in_bytes = 0) { has_texcoords = true;  set_attribute_array(ctx, ref_prog().get_attribute_location(ctx, "texcoord"), texcoords, nr_elements, stride_in_bytes); }
 			/// template method to set the texcoord attribute from a vertex buffer object, the element type must be given as explicit template parameter
-			void set_texcoord_array(const context& ctx, type_descriptor element_type, const vertex_buffer& vbo, size_t offset_in_bytes, size_t nr_elements, size_t stride_in_bytes = 0);
+			void set_texcoord_array(const context& ctx, type_descriptor element_type, const vertex_buffer& vbo, size_t offset_in_bytes, size_t nr_elements, unsigned stride_in_bytes = 0);
 			/// template method to set the texcoord attribute from a vertex buffer object, the element type must be given as explicit template parameter
 			template <typename ElementType>
-			void set_texcoord_array(const context& ctx, const vertex_buffer& vbo, size_t offset_in_bytes, size_t nr_elements, size_t stride_in_bytes = 0) { set_texcoord_array(ctx, type_descriptor(element_descriptor_traits<T>::get_type_descriptor(T()), true), vbo, offset_in_bytes, nr_elements, stride_in_bytes); }
+			void set_texcoord_array(const context& ctx, const vertex_buffer& vbo, size_t offset_in_bytes, size_t nr_elements, unsigned stride_in_bytes = 0) { set_texcoord_array(ctx, type_descriptor(element_descriptor_traits<T>::get_type_descriptor(T()), true), vbo, offset_in_bytes, nr_elements, stride_in_bytes); }
 		};
 	}
 }
